+++ src/_queries/crew/crew-list-queries.tsx
@@ -1,0 +1,25 @@
+import { getCrewList } from '@/src/_apis/crew/crew-list';
+import { ConditionTypes, MainCrewListResponse } from '@/src/types/crew-card';
+
+export function useGetCrewListQuery(condition: ConditionTypes) {
+  return {
+    queryKey: [
+      condition.keyword,
+      condition.mainLocation,
+      condition.mainCategory,
+      condition.subCategory,
+      condition.sortType,
+    ],
+    queryFn: ({ pageParam = 0 }) =>
+      getCrewList(condition, { page: pageParam, size: 6, sort: [condition.sortType] }).then(
+        (response) => {
+          if (response === undefined) {
+            throw new Error('Response is undefined');
+          }
+          return response;
+        },
+      ),
+    getNextPageParam: (lastPage: MainCrewListResponse, allPages: MainCrewListResponse[]) =>
+      lastPage.hasNext ? allPages.length : undefined,
+  };
+}
